Move more render specific functions to render-biblio
